distrib_continuous

Function to compare means or medians in different groups from complex survey data. It produces a table, a graphic and a statistical test.

Usage

distrib_continuous(
  data,
  quanti_exp,
  type = "median",
  facet = NULL,
  quantiles = seq(0.1, 0.9, 0.1),
  filter_exp = NULL,
  ...,
  na.rm.facet = T,
  limits = NULL,
  bw = 1,
  resolution = 1024,
  show_ci_lines = T,
  show_ci_area = F,
  show_center = T,
  show_segments = F,
  show_n = FALSE,
  show_value = TRUE,
  show_lab = TRUE,
  digits = 0,
  unit = "",
  dec = ",",
  pal = c("#c94d33", "#ffe3bf"),
  border = NA,
  font = "Roboto",
  title = NULL,
  subtitle = NULL,
  xlab = NULL,
  ylab = NULL,
  caption = NULL,
  export_path = NULL
)

Arguments

data

A dataframe or an object from the survey package or an object from the srvyr package.

quanti_exp

An expression that define the variable from which the mean/median is computed.

type

"mean" to compute mean by group ; "median" to compute median by group.

facet

A variable defining the faceting group.

filter_exp

An expression that filters the data, preserving the design.

...

All options possible in as_survey_design in srvyr package.

na.rm.facet

TRUE if you want to remove observations with NA on the group variable or NA on the facet variable. FALSE if you want to create a group with the NA value for the group variable and a facet with the NA value for the facet variable. NA in the variables included in prop_exp are not affected in this argument. All the observation with a NA in the variables included in prop_exp are excluded.

show_n

TRUE if you want to show on the graphic the number of individuals in the sample in each group. FALSE if you do not want to show this number. Default is FALSE.

show_value

TRUE if you want to show the mean/median of each group on the graphic. FALSE if you do not want to show the mean/median. Default is TRUE.

show_lab

TRUE if you want to show axes, titles and caption labels. FALSE if you do not want to show any label on axes and titles. Default is TRUE.

digits

Numbers of digits showed on the value labels on the graphic. Default is 0.

unit

Unit showed on the graphic. Default is no unit.

dec

Decimal mark shown on the graphic. Default is ",".

font

Font used in the graphic. Available fonts, included in the package itself, are "Roboto", "Montserrat" and "Gotham Narrow". Default is "Roboto".

title

Title of the graphic.

subtitle

Subtitle of the graphic.

xlab

X label on the graphic. As coord_flip() is used in the graphic, xlab refers to the X label on the graphic, after the coord_flip(), and not to the x variable in the data. If xlab = NULL, X label on the graphic will be "Moyenne : " + quanti_exp or "Médianne : " + quanti_exp. To show no X label, use xlab = "".

ylab

Y label on the graphic. As coord_flip() is used in the graphic, ylab refers to the Y label on the graphic, after the coord_flip(), and not to the y variable in the data. If ylab = NULL, Y label on the graphic will be group. To show no Y label, use ylab = "".

caption

Caption of the graphic.

export_path

Path to export the results in an xlsx file. The file includes three sheets : the table, the graphic and the statistical test.

show_ci

TRUE if you want to show the error bars on the graphic. FALSE if you do not want to show the error bars. Default is TRUE.

Value

A list that contains a table, a graphic and a statistical test