distrib_group_discrete

Usage

distrib_group_discrete(
  data,
  group,
  quali_var,
  facet = NULL,
  filter_exp = NULL,
  ...,
  na.rm.group = T,
  prop_method = "beta",
  show_value = TRUE,
  show_lab = TRUE,
  scale = 100,
  digits = 0,
  unit = "",
  dec = ",",
  pretty_pal = "Hokusai1",
  direction = 1,
  dodge = 0.9,
  font = "Roboto",
  wrap_width_y = 25,
  wrap_width_leg = 25,
  legend_ncol = 4,
  title = NULL,
  subtitle = NULL,
  xlab = NULL,
  ylab = NULL,
  legend_lab = NULL,
  caption = NULL,
  export_path = NULL
)

distrib_group_d(...)

Arguments

data

A dataframe or an object from the survey package or an object from the srvyr package.

group

A variable defining groups to be compared.

quali_var

The discrete variable that is described in the different groups.

facet

A variable defining the faceting group.

filter_exp

An expression that filters the data, preserving the design.

...

All options possible in as_survey_design in srvyr package.

na.rm.group

TRUE if you want to remove the NAs in quali_var, group and facet. FALSE if you want to create NA categories for quali_var, group and facet. Default is TRUE.

prop_method

Type of proportion method used to compute confidence intervals. See svyciprop in survey package for details. Default is the beta method.

show_value

TRUE if you want to show the proportion in each category of each group on the graphic. FALSE if you do not want to show the proportions. Proportions of 2

show_labTRUE if you want to show axes, titles, caption and legend labels. FALSE if you do not want to show any label on axes, titles, caption and legend. Default is TRUE.

scaleDenominator of the proportion. Default is 100 to interprets numbers as percentages.

digitsNumbers of digits showed on the values labels on the graphic. Default is 0.

unitUnit showed in the graphic. Default is no unit.

decDecimal mark shown on the graphic. Default is ",".

pretty_palColor palette used on the graphic. The palettes from the packages MetBrewer, MoMAColors and PrettyCols are available.

directionDirection of the palette color. Default is 1. The opposite direction is -1.

dodgeWidth of the bar, between 0 and 1. Default is 0.9.

fontFont used in the graphic. Available fonts, included in the package itself, are "Roboto", "Montserrat" and "Gotham Narrow". Default is "Roboto".

wrap_width_yNumber of characters before going to the line for the labels of the groups. Default is 25.

wrap_width_legNumber of characters before going to the line for the labels of the categories of quali_var. Default is 25.

legend_ncolNumber of colomns in the legend. Default is 4.

titleTitle of the graphic.

subtitleSubtitle of the graphic.

xlabX label on the graphic. As coord_flip() is used in the graphic, xlab refers to the x label on the graphic, after the coord_flip(), and not to the x variable in the data. If xlab = NULL, the X label on the graphic wil be Distribution : " + quali_var. To show no X label, use xlab = "".

ylabY label on the graphic. As coord_flip() is used in the graphic, ylab refers to the y label on the graphic, after the coord_flip(), and not to the y variable in the data. If ylab = NULL, Y label on the graphic will be group. To show no Y label, use ylab = "".

legend_labLegend (fill) label on the graphic. If legend_lab = NULL, legend label on the graphic will be quali_var. To show no legend label, use legend_lab = "".

captionCaption of the graphic.

export_pathPath to export the results in an xlsx file. The file includes three sheets : the table, the graphic and the statistical test.

A list that contains a table, a graphic and a statistical test

Function describe the distribution of a discrete variable in different groups. It can use complex survey data. It produces a table, a graphic and a statistical test.

# Loading of data data(eusilc, package = "laeken")# Recoding eusilc$pl030 into eusilc$pl030_rec eusilc$pl030_rec <- NA eusilc$pl030_rec[eusilc$pl030 == "1"] <- "Working full time" eusilc$pl030_rec[eusilc$pl030 == "2"] <- "Working part time" eusilc$pl030_rec[eusilc$pl030 == "3"] <- "Unemployed" eusilc$pl030_rec[eusilc$pl030 == "4"] <- "Student" eusilc$pl030_rec[eusilc$pl030 == "5"] <- "Retired" eusilc$pl030_rec[eusilc$pl030 == "6"] <- "Permanently disabled" eusilc$pl030_rec[eusilc$pl030 == "7"] <- "Fulfilling domestic tasks"# Computation, taking sample design into account eusilc_dist_d <- distrib_group_d( eusilc, group = pb220a, quali_var = pl030_rec, strata = db040, ids = db030, weight = rb050, title = "Distribution of socio-economic status according to nationality", subtitle = "Example with austrian SILC data from 'laeken' package" ) eusilc_dist_d$graph eusilc_dist_d$tab