make_surface

make_surface

Usage

make_surface(
  tab,
  var,
  value,
  error_low = NULL,
  error_upp = NULL,
  facet = NULL,
  pvalue = NULL,
  reorder = F,
  compare = F,
  space = NULL,
  position = "mid",
  show_ci = TRUE,
  name_total = "Total",
  digits = 0,
  unit = NULL,
  pal = "Kandinsky",
  direction = 1,
  bg = "snow2",
  font = "Roboto",
  wrap_width_lab = 20,
  title = NULL,
  subtitle = NULL,
  caption = NULL
)

Arguments

tab

dataframe with the variables to be ploted.

var

The variable in tab with the labels of the indicators to be ploted.

value

The variable in tab with the values of the indicator to be ploted.

error_low

The variable in tab that is the lower bound of the confidence interval. If either error_low or error_upp is NULL error bars are not shown on the graphic.

error_upp

The variable in tab that is the upper bound of the confidence interval. If either error_low or error_upp is NULL error bars are not shown on the graphic.

facet

A variable in tab defining the faceting group, if applicable. Default is NULL.

pvalue

The p-value to show in the caption. It can a numeric value or the pvalue object from a statsistical test.

reorder

TRUE if you want to reorder the values. NA value is not included in the reorder.

compare

TRUE to display a square representing the smallest value. When facets are enabled, this is the smallest value per facet category.

space

The space between the squares. The unit is that of the indicator.

position

The position of the squares: "mid" for center alignment, "bottom" for bottom alignment.

show_ci

TRUE if you want to show the CI on the graphic. The bounds of the confidence intervals are displayed as dotted squares around the result. FALSE if you do not want to show them. Default is TRUE.

name_total

Name of the var label that may contain the total. When indicated, it is not displayed on the graph.

digits

Numbers of digits showed on the values labels on the graphic. Default is 0.

unit

The unit showd on the plot. Default is percent.

pal

Color palette used on the graphic. The palettes from the packages MetBrewer, MoMAColors and PrettyCols are available.

direction

Direction of the palette color. Default is 1. The opposite direction is -1.

bg

Color of the background.

font

Font used in the graphic. See load_and_active_fonts() for available fonts.

wrap_width_lab

Number of characters before going to the line for the labels of the categories of var. Default is 20.

title

Title of the graphic.

subtitle

Subtitle of the graphic.

caption

Caption of the graphic.

Value

A ggplot graphic.