distrib_discrete

Usage

distrib_discrete(
  data,
  quali_var,
  facet = NULL,
  filter_exp = NULL,
  ...,
  na.rm.group = T,
  probs = NULL,
  prop_method = "beta",
  reorder = FALSE,
  show_ci = TRUE,
  show_n = FALSE,
  show_value = TRUE,
  show_lab = TRUE,
  scale = 100,
  digits = 0,
  unit = "%",
  dec = ",",
  fill = "sienna2",
  dodge = 0.9,
  font = "Roboto",
  wrap_width_y = 25,
  title = NULL,
  subtitle = NULL,
  xlab = NULL,
  ylab = NULL,
  caption = NULL,
  export_path = NULL
)

distrib_d(...)

Arguments

data

A dataframe or an object from the survey package or an object from the srvyr package.

quali_var

The discrete variable that is described.

facet

A variable defining the faceting group.

filter_exp

An expression that filters the data, preserving the design.

...

All options possible in as_survey_design in srvyr package.

na.rm.group

TRUE if you want to remove observations with NA in quali_var and observations with NA on facet if applicable. FALSE if you want to create FALSE if you want to create a NA modality in quali_var and a facet with the NA value in facet if applicable. Default is TRUE.

probs

Vector of probabilities for H0 of the statistical test, in the correct order (will be rescaled to sum to 1). If probs = NULL, no statistical test is performed. Default is NULL.

prop_method

Type of proportion method to use to compute confidence intervals. See svyciprop in survey package for details. Default is the beta method.

reorder

TRUE if you want to reorder the categories according to their proportion. NA value, in case if na.rm.group = FALSE, is not included in the reorder. Default is FALSE.

show_ci

TRUE if you want to show the error bars on the graphic. FALSE if you do not want to show the error bars. Default is TRUE.

show_n

TRUE if you want to show on the graphic the number of individuals in the sample in each modality of quali_var. FALSE if you do not want to show this number. Default is FALSE.

show_value

TRUE if you want to show the proportion of each category on the graphic. FALSE if you do not want to show the proportion. Default is TRUE.

show_lab

TRUE if you want to show axes, titles and caption labels. FALSE if you do not want to show any label on axes and titles. Default is TRUE.

scale

Denominator of the proportion. Default is 100 to interprets numbers as percentages.

digits

Numbers of digits showed on the values labels on the graphic. Default is 0.

unit

Unit showed in the graphic. Default is "

decDecimal mark shown on the graphic. Default is ",".

fillColour of the bars. NA bar, in case if na.rm.group = FALSE, and total bar are always in grey.

dodgeWidth of the bar, between 0 and 1. Default is 0.9.

fontFont used in the graphic. Available fonts, included in the package itself, are "Roboto", "Montserrat" and "Gotham Narrow". Default is "Roboto".

wrap_width_yNumber of characters before going to the line in the labels of the categories. Default is 25.

titleTitle of the graphic.

subtitleSubtitle of the graphic.

xlabX label on the graphic. As coord_flip() is used in the graphic, xlab refers to the X label on the graphic, after the coord_flip(), and not to the x variable in the data. If xlab = NULL, X label on the graphic will be "Distribution (total : 100ylabY label on the graphic. As coord_flip() is used in the graphic, ylab refers to the Y label on the graphic, after the coord_flip(), and not to the Y variable in the data. If ylab = NULL, Y label on the graphic will be quali_var. To show no Y label, use ylab = "".captionCaption in the graphic.export_pathPath to export the results in an xlsx file. The file includes three sheets : the table, the graphic and the statistical test (if probs is not NULL). A list that contains a table, a graphic and a statistical test Function describe the distribution of a discrete variable from complex survey data. It produces a table and a graphic. # Loading of data data(eusilc, package = "laeken")# Recoding eusilc$pl030 into eusilc$pl030_rec eusilc$pl030_rec <- NA eusilc$pl030_rec[eusilc$pl030 == "1"] <- "Working full time" eusilc$pl030_rec[eusilc$pl030 == "2"] <- "Working part time" eusilc$pl030_rec[eusilc$pl030 == "3"] <- "Unemployed" eusilc$pl030_rec[eusilc$pl030 == "4"] <- "Student" eusilc$pl030_rec[eusilc$pl030 == "5"] <- "Retired" eusilc$pl030_rec[eusilc$pl030 == "6"] <- "Permanently disabled" eusilc$pl030_rec[eusilc$pl030 == "7"] <- "Fulfilling domestic tasks"# Computation, taking sample design into account eusilc_dist_group_d <- distrib_d( eusilc, pl030_rec, strata = db040, ids = db030, weight = rb050, title = "Distribution of socio-economic status", subtitle = "Example with austrian SILC data from 'laeken' package" ) eusilc_dist_group_d$graph eusilc_dist_group_d$tab