esth_graph

Function to construct a graphic following the aestetics of the other function function of this package from a table

Usage

esth_graph(
  tab,
  var,
  value,
  facet = NULL,
  reorder = F,
  error_low = NULL,
  error_upp = NULL,
  n_var = NULL,
  show_value = TRUE,
  name_total = NULL,
  scale = 1,
  digits = 2,
  unit = "",
  dec = ",",
  fill = "indianred4",
  dodge = 0.9,
  font = "Roboto",
  wrap_width_y = 25,
  title = NULL,
  subtitle = NULL,
  xlab = NULL,
  ylab = NULL,
  caption = NULL
)

Arguments

tab

Table with the variables to be ploted.

var

The variable in tab with the labels of the indicators to be ploted.

value

The variable in tab with the indicators to be ploted.

facet

A variable in tab defining the faceting group, if applicable. Default is NULL.

reorder

TRUE if you want to reorder var according to value. FALSE if you do not want to reorder. Default is FALSE.

error_low

The variable in tab that is the lower bound of the confidence interval. If either error_low or error_upp is NULL error bars are not shown on the graphic.

error_upp

The variable in tab that is the upper bound of the confidence interval. If either error_low or error_upp is NULL error bars are not shown on the graphic.

n_var

The variable in tab containing the number of observation per for each indicator. Default is NULL, not showing the number of observation on the plot.

show_value

TRUE if you want to show the values of value on the graphic. FALSE if you do not want to show the proportion. Default is TRUE.

name_total

Name of the total bar on the graphic.

scale

Denominator of the proportion. Default is 100 to interprets numbers as percentages.

digits

Numbers of digits showed on the values labels on the graphic. Default is 0.

unit

The unit showd on the plot. Default is percent.

dec

Decimal mark shown on the graphic. Default is ","

fill

Colour of the bars.

dodge

Width of the bar, between 0 and 1.

font

Font used in the graphic. Available fonts, included in the package itself, are "Roboto", "Montserrat" and "Gotham Narrow". Default is "Roboto".

wrap_width_y

Number of characters before before going to the line. Applies to the labels var. Default is 25.

title

Title of the graphic.

subtitle

Subtitle of the graphic.

xlab

X label on the graphic. As coord_flip() is used in the graphic, xlab refers to the x label on the graphic, after the coord_flip(), and not to var in tab.

ylab

Y label on the graphic. As coord_flip() is used in the graphic, ylab refers to the y label on the graphic, after the coord_flip(), and not to value in tab.

caption

Caption of the graphic.